small updates on unti test

+++ src/main/java/com/revature/Track2gether/controller/ProductionController.java
@@ -1,6 +1,7 @@
 package com.revature.Track2gether.controller;
 
 
+import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.revature.Track2gether.dto.*;
 
@@ -13,6 +14,8 @@ import com.revature.Track2gether.service.AuthenticationService;
 import com.revature.Track2gether.service.JwtService;
 import com.revature.Track2gether.service.TransactionService;
 import com.revature.Track2gether.service.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpHeaders;
@@ -21,10 +24,12 @@ import org.springframework.web.bind.annotation.*;
 import io.jsonwebtoken.MalformedJwtException;
 
 
+import javax.persistence.EntityNotFoundException;
 import javax.security.auth.login.FailedLoginException;
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.time.format.DateTimeParseException;
 import java.util.*;
 
 @RestController
@@ -44,6 +49,8 @@ public class ProductionController {
     @Autowired
     private TransactionService transactionservice;
 
+    Logger logger= LoggerFactory.getLogger(ProductionController.class);
+
     @Autowired
     private CategoryRepository catrepo;
     Transactiondto dto = new Transactiondto();
@@ -55,6 +62,7 @@ public class ProductionController {
     @PostMapping("/login")
     public ResponseEntity<?> login(@RequestBody LoginDTO dto) throws JsonProcessingException {
         try {
+            // logger.info("Login functionality...");
             Users users = authService.login(dto.getEmail(), dto.getPassword());
 
             String jwt = jwtService.createJwt(users);
@@ -112,66 +120,130 @@ public class ProductionController {
     }
 
     @PostMapping("/users/{userid}/transaction")
-    public ResponseEntity<?> addTransaction(@PathVariable("userid") String userid, @RequestBody Transactiondto dto) throws ParseException {
-        Transaction transadd = new Transaction();
-        Users user = userservice.getUserById(Integer.parseInt(userid));
-        Category category = catrepo.getById(dto.getCategoryid());
-        Date dt = df.parse(dto.getDate());
-        transadd.setAmount(dto.getAmount());
-        transadd.setDate(dt);
-        transadd.setDescription(dto.getDescription());
-        transadd.setUser(user);
-        transadd.setCategory(category);
-        transadd.setShared(dto.isShared());
-        dto = transactionservice.addTransaction(transadd);
-        return ResponseEntity.ok(dto);
-    }
+    public ResponseEntity<?> addTransaction(@RequestHeader("Authorization") String headerValue,@PathVariable("userid") String userid, @RequestBody Transactiondto dto) throws ParseException, JsonProcessingException, BadParameterException {
+        try{
+            logger.info("Adding transaction...");
+            String jwt = headerValue.split(" ")[1];
+            Transaction transadd = new Transaction();
+
+            Users user = userservice.getUserById(Integer.parseInt(userid));
+            Category category = catrepo.getById(dto.getCategoryid());
+            Date dt = df.parse(dto.getDate());
+            transadd.setAmount(dto.getAmount());
+            transadd.setDate(dt);
+            transadd.setDescription(dto.getDescription());
+            transadd.setUser(user);
+            transadd.setCategory(category);
+            transadd.setShared(dto.isShared());
+            try {
+                UserJwtDto userdto = jwtService.parseJwt(jwt);
+                if(userdto.getUserId()==user.getId())
+                {    dto = transactionservice.addTransaction(transadd);
+                    System.out.println(userdto);
+                    return ResponseEntity.ok(dto);}
+                else{
+                    return ResponseEntity.status(401).body("You are not allowed to access this endpoint ");
+                }
+            } catch (JsonProcessingException e) {
+                return ResponseEntity.status(401).body(e.getMessage());
+
+            }}catch (Exception e) {
+            return ResponseEntity.status(401).body(e.getMessage()+"Please validate input");
+
+        }}
 
     /*___________________________________*/
     @GetMapping("/users/{userid}/transactions")
-    public ResponseEntity<?> getTransactionByUserId(@PathVariable("userid") String userid,
-                                                    @RequestParam("transtype") Optional<Integer> transtype) {
-        Transaction transadd = new Transaction();
-        Users user = userservice.getUserById(Integer.parseInt(userid));
-        List<Transactiondto> responses = new ArrayList<Transactiondto>();
-        if(transtype.isPresent()){
-            responses= transactionservice.findByTransactiontype(user, transtype.get());
-        }else{
-            responses = transactionservice.findByUser(user);
-        }
-        return ResponseEntity.ok(responses);
+    public ResponseEntity<?> getTransactionByUserId(@RequestHeader("Authorization")String headerValue, @PathVariable("userid") String userid,
+                                                    @RequestParam("transtype") Optional<Integer> transtype) throws BadParameterException {
+        try{
+            Transaction transadd = new Transaction();
+            Users user = userservice.getUserById(Integer.parseInt(userid));
+            String jwt = headerValue.split(" ")[1];
+            try {
+                UserJwtDto userdto = jwtService.parseJwt(jwt);
+                if(userdto.getUserId()==user.getId()) {
+                    List<Transactiondto> responses = new ArrayList<Transactiondto>();
+                    if (transtype.isPresent()) {
+                        logger.info("Get all transactions of a user by transactiontype...");
+                        responses = transactionservice.findByTransactiontype(user, transtype.get());
+                    } else {
+                        logger.info("Get all transactions of a user...");
+                        responses = transactionservice.findByUser(user);
+                    }
+                    return ResponseEntity.ok(responses);
+                } else{
+                    return ResponseEntity.status(401).body("You are not allowed to access this endpoint ");
+                }
+            }catch (JsonProcessingException e) {
+                return ResponseEntity.status(401).body(e.getMessage());
+
+            }}catch (Exception e) {
+            return ResponseEntity.status(401).body(e.getMessage()+"Please validate input");}
+
     }
 
     @GetMapping("/users/{userid}/transactions/filterby")
-    public ResponseEntity<?> getTransactionByTranstype(@PathVariable("userid") String userid,
+    public ResponseEntity<?> getTransactionByTranstype(@RequestHeader("Authorization")String headerValue ,@PathVariable("userid") String userid,
                                                        @RequestParam("year") int year,
-                                                       @RequestParam("month") int month){
-        Transaction transadd = new Transaction();
-        Users user = userservice.getUserById(Integer.parseInt(userid));
-        List<Transactiondto> responses = new ArrayList<Transactiondto>();
-        if(year!=0 && month!=0){
-            responses = transactionservice.findByTransactions( year , month);
-        }
+                                                       @RequestParam("month") int month) throws BadParameterException {
 
-        return ResponseEntity.ok(responses);
+        try{
+            logger.info("Get all transactions of a user by month and year...");
+            Transaction transadd = new Transaction();
+            Users user = userservice.getUserById(Integer.parseInt(userid));
+            String jwt = headerValue.split(" ")[1];
+            try {
+                UserJwtDto userdto = jwtService.parseJwt(jwt);
+                if(userdto.getUserId()==user.getId()) {
+                    List<Transactiondto> responses = new ArrayList<Transactiondto>();
+                    if(year!=0 && month!=0){
+                        responses = transactionservice.findByTransactions( year , month);
+                    }
+
+                    return ResponseEntity.ok(responses);}else{
+                    return ResponseEntity.status(401).body("You are not allowed to access this endpoint ");
+                }
+            }catch (JsonProcessingException e) {
+                return ResponseEntity.status(401).body(e.getMessage());
+
+            }}catch (Exception e) {
+            return ResponseEntity.status(401).body(e.getMessage()+"Please validate input");}
     }
 
     /*___________________________________*/
     @PutMapping("/users/{userid}/transaction/{id}")
-    public ResponseEntity<?> updateTransaction(@PathVariable("userid") String userid,@PathVariable("id") String id,@RequestBody Transactiondto dto) throws ParseException {
-        Transaction transadd = new Transaction();
-        Users user = userservice.getUserById(Integer.parseInt(userid));
-        Category category = catrepo.getById(dto.getCategoryid());
-        transadd.setId(Integer.parseInt(id));
-        Date dt = df.parse(dto.getDate());
-        transadd.setAmount(dto.getAmount());
-        transadd.setDate(dt);
-        transadd.setDescription(dto.getDescription());
-        transadd.setUser(user);
-        transadd.setCategory(category);
-        transadd.setShared(dto.isShared());
-        Transactiondto newtrans = transactionservice.updateTransaction(transadd);
-        return ResponseEntity.ok(newtrans);
+    public ResponseEntity<?> updateTransaction(@RequestHeader("Authorization")String headerValue,@PathVariable("userid") String userid,@PathVariable("id") String id,@RequestBody Transactiondto dto) throws ParseException, BadParameterException {
+        try{
+            logger.info("Update transaction of a user...");
+            Transaction transadd = new Transaction();
+            Users user = userservice.getUserById(Integer.parseInt(userid));
+            Category category = catrepo.getById(dto.getCategoryid());
+            transadd.setId(Integer.parseInt(id));
+            Date dt = df.parse(dto.getDate());
+            transadd.setAmount(dto.getAmount());
+            transadd.setDate(dt);
+            transadd.setDescription(dto.getDescription());
+            transadd.setUser(user);
+            transadd.setCategory(category);
+            transadd.setShared(dto.isShared());
+            String jwt = headerValue.split(" ")[1];
+            try {
+                UserJwtDto userdto = jwtService.parseJwt(jwt);
+                if(userdto.getUserId()==user.getId()) {
+                    Transactiondto newtrans = transactionservice.updateTransaction(transadd);
+                    return ResponseEntity.ok(newtrans);
+                }else{
+                    return ResponseEntity.status(401).body("You are not allowed to access this endpoint ");
+                }
+
+            }catch (JsonProcessingException e) {
+                return ResponseEntity.status(401).body(e.getMessage());
+
+            }}catch (Exception e) {
+            return ResponseEntity.status(401).body(e.getMessage()+"Please validate input");}
+
+
 
     }
     /*___________________________________*/
@@ -179,19 +251,35 @@ public class ProductionController {
     public ResponseEntity<?> getAllCategory(@RequestParam("type") int transid)
 
     {
+        logger.info("Get all Category...");
         List<Categorydto> cdto = transactionservice.findByCategoryBytranstype(transid);
         return ResponseEntity.ok(cdto);
     }
 
 
     @DeleteMapping("/users/{userid}/transaction/{id}")
-    public ResponseEntity<?> deleteTransaction(@PathVariable("userid") String userid,@PathVariable("id") String id)
-    {
-        int transid=Integer.parseInt(id);
-        Users user = userservice.getUserById(Integer.parseInt(userid));
-        transactionservice.deleteTransactionById(transid);
-        return ResponseEntity.ok().build();
-    }
+    public ResponseEntity<?> deleteTransaction(@RequestHeader("Authorization")String headerValue,@PathVariable("userid") String userid,@PathVariable("id") String id) throws EntityNotFoundException, BadParameterException {
+        try {
+            logger.info("Deleting transaction of a user...");
+            int transid = Integer.parseInt(id);
+            Users user = userservice.getUserById(Integer.parseInt(userid));
+            String jwt = headerValue.split(" ")[1];
+            try {
+                UserJwtDto userdto = jwtService.parseJwt(jwt);
+                if (userdto.getUserId() == user.getId()) {
+                    transactionservice.deleteTransactionById(transid);
+                    return ResponseEntity.ok().build();
+                } else {
+                    return ResponseEntity.status(401).body("You are not allowed to access this endpoint ");
+                }
 
+            } catch (JsonProcessingException e) {
+                return ResponseEntity.status(401).body(e.getMessage());
+
+            }
+        } catch (Exception e) {
+            return ResponseEntity.status(401).body(e.getMessage() + "Please validate input");
+        }
+    }
 
 }
